emacs: Set leader key combinations in general.el's :config
Instead of in the :config of every package for which the keys are
relevant. This did not work correctly, as the use-package forms are
sorted alphabetically, but the keys must be set after general.el's
config was executed.

Another solution may be to define a custom general definer, and use it
with :general in the use-package forms, but I haven't tried.