vterm: Fix C-n and C-p not working
I thought this was a problem of evil-mode and/or evil-collection, but
it was my own fault: in a non-Emacs terminal emulator I use vi
keybindings in bash and hence C-n and C-p do not work. In Emacs,
because I use evil-mode, ESC does not exit the shell's vi insert mode,
but the Emacs buffer's insert mode. If the shell is started from
within Emacs, I thus do not want to use vi keybindings.