mopidy-mpd: Use my fork of mopidy-mpd
The fork supports using multiple MPD tags in the case of e.g. multiple
artists of a track. I have opened a pull request to merge this
upstream, but it is not yet merged. In the meantime (i.e., before the
changes land in Nixpkgs (unstable)), use this source override.