notmuch-update: Do not require a function call to setup hooks
Since notmuch-update requires notmuch, it will load after notmuch has
loaded. I have setup notmuch to autoload when the notmuch command is
used, so notmuch-update would only load after notmuch started. It
would then add the required hook to setup the timer too late. With the
`notmuch-update-start' function, it is not required to call the
`notmuch-update-add-hook' function to setup the required hooks.